Session moved to session_types and renamed BaseSession.
