+++ robot-server/robot_server/service/session/session_types/__init__.py
@@ -1,0 +1,2 @@
+from .null_session import NullBaseSession
+from .check_session import CheckBaseSession
