+++ robot-server/robot_server/service/session/session_types/null_session.py
@@ -1,16 +1,13 @@
-from functools import lru_cache
-from typing import Optional
-
 from opentrons.calibration.check.models import SessionType
 
 from robot_server.service.session import models
 from robot_server.service.session.command_execution import CommandQueue,\
     CommandExecutor
 from robot_server.service.session.configuration import SessionConfiguration
-from robot_server.service.session.session import Session, SessionMetaData
+from robot_server.service.session.session_types.base_session import BaseSession, SessionMetaData
 
 
-class NullSession(Session):
+class NullBaseSession(BaseSession):
     """A session that does nothing."""
 
     def __init__(self,
